requirements-test.txt
-r requirements.txt
pytest-django==4.9.0
mypy==1.11.0
flake8==7.1.1
django-stubs==5.1.0
djangorestframework-stubs[compatible-mypy]==3.15.1
freezegun==1.5.1
pytest-asyncio==0.24.0
syrupy==4.8.0
